refactored code to divide in several import files
